Backend/requirements.txt
Flask
Flask-Cors
Flask-Limiter
Celery
redis
pymongo
scikit-learn
joblib
flask-pymongo
python-dotenv
pandas
python-dateutil



